validation: add nLockTime and MTP validation (BIP113) (#24)
* validation: add nLockTime and MTP validation (BIP113)

Implement transaction locktime validation for both block and mempool acceptance:

- Add locktime constants (LOCKTIME_THRESHOLD, SEQUENCE_FINAL, MTP_BLOCKS)
- Add helper functions: is_tx_final_by_sequence(), compute_mtp(), check_locktime()
- Add LocktimeNotSatisfied error variant
- Update validate_block_basic() to check locktimes with MTP parameter
- Add locktime check to mempool acceptance (add_to_mempool)
- Add Node::compute_mtp() to calculate Median Time Past from last 11 blocks
- Add comprehensive test suite for locktime validation

Locktime rules implemented:
- nLockTime < 500,000,000 -> block height comparison
- nLockTime >= 500,000,000 -> Unix timestamp comparison against MTP
- If ALL inputs have sequence=0xffffffff, transaction is final (locktime ignored)



* fix: address clippy warnings for CI

- Collapse nested if-let statements in compute_mtp
- Remove unnecessary u32 cast in test file



---------

+++ tests/node_phase1b1_p2p_multi.rs
@@ -67,7 +67,7 @@ fn start_bad_peer(magic: [u8; 4]) -> String {
         let mut header = Vec::new();
         header.extend_from_slice(&magic);
         header.extend_from_slice(b"junk\0\0\0\0\0\0\0\0");
-        header.extend_from_slice(&((MAX_MESSAGE_BYTES + 1) as u32).to_le_bytes());
+        header.extend_from_slice(&(MAX_MESSAGE_BYTES + 1).to_le_bytes());
         header.extend_from_slice(&[0u8; 4]);
         header.extend_from_slice(&payload[..16]); // send partial to exercise read logic
         let _ = stream.write_all(&header);
